Support additional track types
Midi files may contain chunks ("tracks") not of type "MTrk".  For
example, files in Yamaha's XF extended midi format can contain an XFIH
(XF Information Header) chunk and a XFKM (XF Karaoke Message) chunk in
addition to the MTrk data.

This commit also adds validation checks for premature end of data, and
for tracks that do not end with an End of Track event.